rename viscous_burgers -> burgers_viscous (#267)
this is consistent with incompressible and also allows it to
appear next to burgers in a listing (alphabetically)
