+++ docs/source/pyro.advection_fv4.problems.rst
@@ -16,3 +16,19 @@ pyro.advection\_fv4.problems.smooth module
    :members:
    :undoc-members:
    :show-inheritance:
+
+pyro.advection\_fv4.problems.test module
+----------------------------------------
+
+.. automodule:: pyro.advection_fv4.problems.test
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.advection\_fv4.problems.tophat module
+------------------------------------------
+
+.. automodule:: pyro.advection_fv4.problems.tophat
+   :members:
+   :undoc-members:
+   :show-inheritance:
