+++ docs/source/pyro.burgers.problems.rst
@@ -1,5 +1,5 @@
 pyro.burgers.problems package
-===============================
+=============================
 
 .. automodule:: pyro.burgers.problems
    :members:
@@ -9,15 +9,6 @@ pyro.burgers.problems package
 Submodules
 ----------
 
-pyro.burgers.problems.test module
--------------------------------------
-
-.. automodule:: pyro.burgers.problems.test
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
 pyro.burgers.problems.converge module
 -------------------------------------
 
@@ -26,10 +17,26 @@ pyro.burgers.problems.converge module
    :undoc-members:
    :show-inheritance:
 
+pyro.burgers.problems.test module
+---------------------------------
+
+.. automodule:: pyro.burgers.problems.test
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 pyro.burgers.problems.tophat module
--------------------------------------
+-----------------------------------
 
 .. automodule:: pyro.burgers.problems.tophat
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.burgers.problems.verify module
+-----------------------------------
+
+.. automodule:: pyro.burgers.problems.verify
    :members:
    :undoc-members:
    :show-inheritance:
