+++ docs/source/pyro.burgers_viscous.problems.rst
@@ -1,0 +1,34 @@
+pyro.burgers\_viscous.problems package
+======================================
+
+.. automodule:: pyro.burgers_viscous.problems
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+pyro.burgers\_viscous.problems.converge module
+----------------------------------------------
+
+.. automodule:: pyro.burgers_viscous.problems.converge
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.burgers\_viscous.problems.test module
+------------------------------------------
+
+.. automodule:: pyro.burgers_viscous.problems.test
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.burgers\_viscous.problems.tophat module
+--------------------------------------------
+
+.. automodule:: pyro.burgers_viscous.problems.tophat
+   :members:
+   :undoc-members:
+   :show-inheritance:
