+++ docs/source/pyro.burgers_viscous.rst
@@ -1,0 +1,34 @@
+pyro.burgers\_viscous package
+=============================
+
+.. automodule:: pyro.burgers_viscous
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pyro.burgers_viscous.problems
+
+Submodules
+----------
+
+pyro.burgers\_viscous.interface module
+--------------------------------------
+
+.. automodule:: pyro.burgers_viscous.interface
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.burgers\_viscous.simulation module
+---------------------------------------
+
+.. automodule:: pyro.burgers_viscous.simulation
+   :members:
+   :undoc-members:
+   :show-inheritance:
