+++ docs/source/pyro.multigrid.examples.rst
@@ -1,0 +1,106 @@
+pyro.multigrid.examples package
+===============================
+
+.. automodule:: pyro.multigrid.examples
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+pyro.multigrid.examples.mg\_test\_general\_alphabeta\_only module
+-----------------------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_general_alphabeta_only
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_general\_beta\_only module
+------------------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_general_beta_only
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_general\_constant module
+----------------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_general_constant
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_general\_dirichlet module
+-----------------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_general_dirichlet
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_general\_inhomogeneous module
+---------------------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_general_inhomogeneous
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_simple module
+-----------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_simple
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_vc\_constant module
+-----------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_vc_constant
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_vc\_dirichlet module
+------------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_vc_dirichlet
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_test\_vc\_periodic module
+-----------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_test_vc_periodic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.mg\_vis module
+--------------------------------------
+
+.. automodule:: pyro.multigrid.examples.mg_vis
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.project\_periodic module
+------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.project_periodic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.multigrid.examples.prolong\_restrict\_demo module
+------------------------------------------------------
+
+.. automodule:: pyro.multigrid.examples.prolong_restrict_demo
+   :members:
+   :undoc-members:
+   :show-inheritance:
