+++ docs/source/pyro.rst
@@ -18,7 +18,7 @@ Subpackages
    pyro.advection_rk
    pyro.advection_weno
    pyro.burgers
-   pyro.viscous_burgers
+   pyro.burgers_viscous
    pyro.compressible
    pyro.compressible_fv4
    pyro.compressible_react
@@ -37,10 +37,34 @@ Subpackages
 Submodules
 ----------
 
+pyro.compare module
+-------------------
+
+.. automodule:: pyro.compare
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.mg\_test\_simple module
+----------------------------
+
+.. automodule:: pyro.mg_test_simple
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 pyro.plot module
 ----------------
 
 .. automodule:: pyro.plot
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyro.plotvar module
+-------------------
+
+.. automodule:: pyro.plotvar
    :members:
    :undoc-members:
    :show-inheritance:
