--- docs/source/pyro.viscous_burgers.problems.rst
@@ -1,26 +1,0 @@
-pyro.viscous\_burgers.problems package
-=======================================
-
-.. automodule:: pyro.viscous_burgers.problems
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-pyro.viscous\_burgers.problems.converge module
------------------------------------------------
-
-.. automodule:: pyro.viscous_burgers.problems.converge
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyro.viscous\_burgers.problems.tophat module
---------------------------------------------
-
-.. automodule:: pyro.viscous_burgers.problems.tophat
-   :members:
-   :undoc-members:
-   :show-inheritance:
