--- docs/source/pyro.viscous_burgers.rst
@@ -1,34 +1,0 @@
-pyro.viscous\_burgers package
-==============================
-
-.. automodule:: pyro.viscous_burgers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   pyro.viscous_burgers.problems
-
-Submodules
-----------
-
-pyro.viscous\_burgers.interface module
----------------------------------------
-
-.. automodule:: pyro.viscous_burgers.interface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyro.viscous_burgers.simulation module
----------------------------------------
-
-.. automodule:: pyro.viscous_burgers.simulation
-   :members:
-   :undoc-members:
-   :show-inheritance:
