+++ pyro/burgers_viscous/_defaults
