+++ pyro/burgers_viscous/problems/inputs.smooth
