+++ pyro/burgers_viscous/problems/inputs.tophat
