+++ pyro/burgers_viscous/simulation.py
@@ -1,7 +1,7 @@
 from pyro.burgers import Simulation as burgers_sim
 from pyro.burgers import burgers_interface
+from pyro.burgers_viscous import interface
 from pyro.mesh import reconstruction
-from pyro.viscous_burgers import interface
 
 
 class Simulation(burgers_sim):
