+++ pyro/pyro_sim.py
@@ -18,7 +18,7 @@ valid_solvers = ["advection",
                  "advection_fv4",
                  "advection_weno",
                  "burgers",
-                 "viscous_burgers",
+                 "burgers_viscous",
                  "compressible",
                  "compressible_rk",
                  "compressible_fv4",
